requirements.txt
requests==2.24.0
requests-html==0.10.0
SQLAlchemy==1.3.18
streamlit==1.0.0
requests-cache==0.8.1
numpy==1.20.2
pandas==1.2.4
plotly==5.3.1
tweepy==4.1.0
streamlit-aggrid==0.2.2.post2
mysql-connector-python==8.0.23
timeago==1.0.15